jodell-claw-control-backend/requirements.txt
pyserial == 3.2.1
minimalmodbus == 2.0.1
